recreate E2E tests for all json mappers in Java tests, delete cypress config/tests
