--- core/src/main/resources/org/jobrunr/dashboard/frontend/cypress.config.js
@@ -1,10 +1,0 @@
-const {defineConfig} = require('cypress')
-
-module.exports = defineConfig({
-    e2e: {
-        supportFile: false,
-        specPattern: 'cypress/integration/*.js',
-        testIsolation: false
-    },
-    video: false
-})
