+++ core/src/main/resources/org/jobrunr/dashboard/frontend/package.json
@@ -43,8 +43,6 @@
   "devDependencies": {
     "@rsbuild/core": "^1.1.13",
     "@rsbuild/plugin-react": "^1.1.0",
-    "cypress": "^13.17.0",
-    "cypress-failed-log": "^2.10.0",
     "jest": "^29.7.0",
     "owasp-dependency-check": "^0.0.24"
   }
