+++ tests/e2e-ui/src/test/java/org/jobrunr/tests/e2e/DashboardActionsGsonE2ETest.java
@@ -1,0 +1,16 @@
+package org.jobrunr.tests.e2e;
+
+import org.jobrunr.tests.server.SimpleBackgroundJobServer;
+import org.junit.jupiter.api.BeforeAll;
+
+public class DashboardActionsGsonE2ETest extends AbstractDashboardActionsE2ETest {
+
+    @BeforeAll
+    static void beforeAll() {
+        server = new SimpleBackgroundJobServer()
+                .withGsonMapper()
+                .withPaused();
+        server.start();
+    }
+
+}
